Refactor CIDatabaseTestCase to use a trait instead.
