+++ user_guide_src/source/changelogs/v4.1.2.rst
@@ -23,3 +23,4 @@ Deprecations:
 - Deprecated cookie-related properties of ``Security`` in order to use the ``Cookie`` class.
 - Deprecated cookie-related properties of ``Session`` in order to use the ``Cookie`` class.
 - Deprecated ``Security::isExpired()`` to use the ``Cookie``'s internal expires status.
+- Deprecated ``CIDatabaseTestCase`` to use the ``DatabaseTestTrait`` instead.
